requirements.txt
Django>=4.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv==1.0.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
sqlalchemy>=2.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
python-multipart>=0.0.6
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
isort>=5.12.0
django-allauth>=0.54.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.3.0
docker>=6.1.3
google-generativeai>=0.3.2
gunicorn>=20.1.0
